parse subsystem product information from pci.ids
The PCI subsystem product information is a twice-TAB-indented block that
appears after a PCI product information block. This patch parses that
subsystem block and adds a VendorId field to PCIProductInfo in order to
represent that a subsystem make have a different vendor than the product
it is a subsystem of.