adds GraphicsCard.Nodes field
Adds a Nodes field to GraphicsCard struct that keeps an array of
pointers to Node structs, one for each NUMA node that the graphics card
is affined to. On non-NUMA systems, this will always be an empty array.

Now, I just need to find a NUMA system with multiple GPUs/graphics cards
to test on.... :)

Issue #4