Allow overriding root path/bindmount
If a host filesystem has been bind-mounted to a particular directory
within an application container (say, '/host'), we want to inform `ghw`
to use that bind-mount as its root directory when constructing paths to
certain files like /proc/cpuinfo or /etc/mtab.

This patch adds the ability for the user to set the GHW_CHROOT environ
variable to the value of this alternate root mountpoint, allowing `ghw`
to find important files from within an application container that has
bind-mounted a host filesystem.

Issue #56