remove dep from make test
It is not necessary or desired to run `dep ensure` when running tests.
`dep ensure` should be used to add or update a dependency, but we only
want to do that explicitly when we either need a newer version of a
dependent library or when adding a new dependency.

Add make dep to the travis pipeline since the Travis test VM starts off
with no dependencies pulled into the GOPATH.