allow warnings to be disabled
With this patch, `ghw` now disables printing of warning messages when a
`GHW_DISABLE_WARNINGS` environ variable is set:

```
[jaypipes@uberbox ghw]$ go run cmd/ghwc/main.go memory
WARNING:
Could not determine total physical bytes of memory. This may
be due to the host being a virtual machine or container with no
/var/log/syslog file, or the current user may not have necessary
privileges to read the syslog. We are falling back to setting the
total physical amount of memory to the total usable amount of memory
memory (24GB physical, 24GB usable)
[jaypipes@uberbox ghw]$ GHW_DISABLE_WARNINGS=1 go run cmd/ghwc/main.go
memory
memory (24GB physical, 24GB usable)
```

Issue #110