[windows] look in diskdrive.Caption for "ssd"
The Win32_DiskDrive.MediaType field apparently is not reliable for
having "fixed" or "ssd" in it for anything prior to Windows 8, so the
recommended "solution" to determine SSDs for Windows is to also look in
the Win32_DiskDrive.Caption field for the string "ssd".

We also remove whitespace from the caption/name fields for partitions
and block devices on Windows.

Fixes Issue #177