first pass packagizing ghw
The main directory of `ghw` is getting a little long in the tooth, so
I'm going to be moving things into their own packages. This first pass
puts the (previously private, now public) Context and Option structs
into their own packages and updates all the primary call logic to pass
in a Context object instead of having methods on the private context
struct.