hack: expand snapshots inside docker
GHW gained support to consume snapshot programmatically.
This means there is no need anymore to expand the snapshot before
mount it in docker, which greatly simplifies the
`run-against-snapshot.sh` scripts.

We just need to pass the snapshot and a bunch of environment variables.

Additionally, start document describing how a GHW compliant snapshot
should look like, and why things are the way they are.

Signed-off-by: Francesco Romani <fromani@redhat.com>