snapshot: move cloning+packing logic in the pkg
Up until now, the logic to perform snapshot was
inside the `ghw-snapshot` tool.
This is because the tool predated the inception
of the `snapshot` package.

However, we can now move most of the snapshot-related
logic into the `snapshot` package:
1. enables better testability - we add tests to demonstrate it!
2. enables code reuse (consumers of ghw can create snapshot
   programmatically)
3. enables a significant refactor, getting rid of a lot
   of repetitive code and introducing simpler, more declarative
   and more generic code.

Please note the refactoring intentionally does not include the
block devices cloning code; rationale for this decision is that
adaptation to new generic code isn't trivial (as for other functions),
so we delay this further simplification to a later PR.

Signed-off-by: Francesco Romani <fromani@redhat.com>