warning: support new option to override default
This patch add support for a new option, which allow
the caller code to pass a new Alerter to ghw.
`Alerter` is a dead-simple interface, supporting only a
`Printf` method much like the stdlib `log` package.
ghw will send all the warnings to this method.

We use a so simple interface because we want to generalize a bit
the existing `util.Warn` concept, but is not clear if ghw will benefit
of a more complete logging interface; should that be the case, we
can build on the concept we laid in this patch.

Client code can programmatically override the Alerter, or
users can using the faimiliar `GHW_DISABLE_WARNINGS` environ variable,
which keep functioning as usual.

ghw can be configured to swallow all the warnings or to keep emitting
them on stderr, like it used to do up until this patch.

Signed-off-by: Francesco Romani <fromani@redhat.com>