context: detect and report conflicting options
`WithChroot` and `WithSnapshot` conflicts to each other,
and the users can get surprising results if they try to use them
at the same time.

Now: chroot is expected to be much more popular than snapshots,
being the latter useful mostly for tests or offline debug/troubleshoot,
but still the code should detect and handle these conflicts.

This is made a bit awkward because the current `context` pkg API.
Once we are free to break compatibility, we can make the flow more
linear.

Signed-off-by: Francesco Romani <fromani@redhat.com>