pci: ensure context ordering
the pci package want to use the topology package, and passes
its context down to it.
However, nontrivial initialization is performed in pci.Info.load(),
when snapshots are consumed - most notably in test.

Hence, we need to ensure that `topology.WithContext` is called only
after context.Setup() is called for the pci context.
This is needed because in turn the pci packaged doesn't know (nor it
should) if its context is a root (= top level) context, hence if it
owns the resources which are passed to the dependent packages.

Signed-off-by: Francesco Romani <fromani@redhat.com>