pci: GetDevice: check the cached data first
Add a simple helper to find a device among the `pci.Info.Devices`.
This is needed because `GetDevice` accesses the system every time.

In most cases, besides a tiny performance hit, this is no issue.
When consuming snapshots, however, this may lead to surprising behaviour
because the caller need to ensure the snapshot is available when
`GetDevice` is called, or it will return unpredictable result (!!!).

The wrong code looks like that:

1. PCI info is created supplying snapshot data with automatic management
1.a ghw unpacks the snapshot on a temporary directory.
1.b ghw loads all the info from the snapshot. info.Devices contains
    correct data.
1.c ghw cleans up the snapshot (see:
  https://github.com/jaypipes/ghw/pull/236)
2. the client (test) code calls GetDevice
2.a GetDevice will try to access the unpacked snapshot data, which is
    gone.
2.b GetDevice will fail.

Signed-off-by: Francesco Romani <fromani@redhat.com>