linuxpaths: allow to override mount points
Inside containers, when granting access to the host file systems, is
common practice to bind-mount specific host subtrees on nonstandard
locations, like `/sys` on `/host-sys` or `/proc` on `/host-proc`.

To better cover this use case, `ghw` allows to *programmatically* override
the initial component of filesystems subtrees, allowing to access `sysfs`
(or `procfs` or...) mounted on these non-standard locations.

Signed-off-by: Francesco Romani <fromani@redhat.com>