clonetree_test: declare it is linux specific
The existing clonetree tests are linux specific because
they weere developed on linux. While no intentional
restriction on linux was put in, no real thought to
other OSes was put in either.

So we just make the expectations explicit and we mark these tests
as linux specific until we have the chance to truly generalize them.

Signed-off-by: Francesco Romani <fromani@redhat.com>