test: remove stale test
Remove generic test that needs to be moved in the os-specific
test until the tier-2 platform (mac, win) catch up with the
implementation.

These tests are meant to be in the linux-specific test from
the beginning, they seem to have sneaked in the generic codebase
due to a rebase glitch.

Signed-off-by: Francesco Romani <fromani@redhat.com>