Get partition labels from udev
This mainly reuses the udevInfo method to obtain the udev info for a
given partition in a given disk instead of having the label value always
empty.

This separates the udevInfo method into 2 distinct methods, one for disk
and one for partition.

This also opens the possibility of obtaining more info for partitions,
i.e. the FS of an unmounted partition.

Signed-off-by: Itxaka <igarcia@suse.com>