assume phys mem == usable mem on ARM
`/sys/devices/system/memory` does not exist on linux/arm systems that do
not support memory hotplug. This means that the
`/sys/devices/system/memory/block_size_bytes` file also does not exist
and that is the file that we use to determine the physical RAM
associated with a RAM module affined to a NUMA node.

This commit updates the logic in `pkg/memory.AreaForNode` on linux
platforms to tolerate the error returned by `memoryBlockSizeBytes()`
when `/sys/devices/system/memory/block_size_bytes` does not exist and
set a memory `Area`'s `TotalPhysicalBytes` value to the value of
either the syslog-determined physical memory total or `TotalUsableBytes`.

Issue #336

Signed-off-by: Jay Pipes <jaypipes@gmail.com>