linux: refactor ProcessorCore retrieval
Reworks the way that ProcessorCore structs are determined on Linux by
breaking up the long `processorsGet` function into smaller functions and
using standard maps for looking up Processor and logicalProcessor
structs.

Removes the `ProcessorCore.Index` field entirely, since it was confusing
and did not actually represent anything useful. Added clarity to the
documentation about `ProcessorCore.LogicalProcessors` and how those
numbers are actually the zero-based index of processors on the host and
are sometimes called "thread siblings" and that zero-based index of
logical processor IDs has nothing to do with the core ID.

Fixes Issue #345

Signed-off-by: Jay Pipes <jaypipes@gmail.com>