doc and const cleanup for block package
Ensures proper docstrings for all exported things in the `block`
package and deprecates the (non-Go-idiomatic) ALL_CAPS constants in
favour of CamelCase constants.

Also officially deprecates the `pkg/block.Info.TotalPhysicalBytes` field
in favour of a `pkg/block.Info.TotalSizeBytes` field (had a TODO in
there a while to do this...).

Signed-off-by: Jay Pipes <jaypipes@gmail.com>