use TotalXXX instead of NumXXX
All other fields that contain a total amount of something begin with the
word `Total` except for the `pkg/cpu.Processor.NumCores` and
`NumThreads` fields. This commit adds new fields to `pkg/cpu.Processor`
called `TotalCores` and `TotalHardwareThreads` to bring the names of
these fields in-line with all other amount/total fields.

In addition, adds a `TotalHardwareThreads` field to the `pkg/cpu.Info`
struct and deprecates the less-explicit `TotalThreads` field.

I'd like to remove the now-deprecated `NumThreads`, `NumCores`, and
`TotalThreads` fields before we get to a `v.1.0` series.

Signed-off-by: Jay Pipes <jaypipes@gmail.com>